Created Database Helper Class
The database helper is an object which helps you connect to the database
and perform a particular call in a simple and user friendly way. The
class supports Select, Update, Insert, and Delete calls takes as
argument the database connection string.

Also added references to the project for MySQL client, and
System.Configuration to allow App.Config reading of the connection
string